refactor: refactor tracing
