+++ cmd/linebot/wire_gen.go
@@ -51,9 +51,9 @@ func register(contextContext context.Context) (*bot, func(), error) {
 		return nil, nil, err
 	}
 	conversation := firestore.NewConversation(client)
-	conversationImpl := service.NewConversation(conversation, tracerProvider)
+	conversationImpl := service.NewConversation(conversation)
 	shopping := firestore.NewShopping(conversation)
-	shoppingImpl := service.NewShopping(conversation, shopping, tracerProvider)
+	shoppingImpl := service.NewShopping(conversation, shopping)
 	parser, err := nl.NewParser()
 	if err != nil {
 		cleanup()
@@ -66,7 +66,7 @@ func register(contextContext context.Context) (*bot, func(), error) {
 		cleanup()
 		return nil, nil, err
 	}
-	reminderImpl := service.NewReminder(reminder, schedulerScheduler, tracerProvider)
+	reminderImpl := service.NewReminder(reminder, schedulerScheduler)
 	interactorReminder := interactor.NewReminder(conversationImpl, reminderImpl, messageProviderSet, botImpl, configConfig)
 	weatherWeather, err := weather.NewWeather(configConfig)
 	if err != nil {
@@ -83,7 +83,7 @@ func register(contextContext context.Context) (*bot, func(), error) {
 		cleanup()
 		return nil, nil, err
 	}
-	weatherImpl := service.NewWeather(weatherWeather, weatherImageStore, configConfig, tracerProvider)
+	weatherImpl := service.NewWeather(weatherWeather, weatherImageStore, configConfig)
 	interactorWeather := interactor.NewWeather(weatherImpl, messageProviderSet, botImpl)
 	eventHandler, err := interactor.NewEventHandler(interactorShopping, interactorReminder, interactorWeather, conversationImpl, reminderImpl, messageProviderSet, botImpl, configConfig)
 	if err != nil {
