+++ domain/service/reminder.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"time"
 
-	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 	"golang.org/x/xerrors"
 
@@ -29,23 +28,20 @@ type Reminder interface {
 type ReminderImpl struct {
 	reminder  repository.Reminder
 	scheduler repository.ScheduleSynchronizer
-	tracer    trace.Tracer
 }
 
 func NewReminder(
 	reminder repository.Reminder,
 	scheduler repository.ScheduleSynchronizer,
-	tracerProvider trace.TracerProvider,
 ) *ReminderImpl {
 	return &ReminderImpl{
 		reminder:  reminder,
 		scheduler: scheduler,
-		tracer:    tracerProvider.Tracer("github.com/ww24/linebot/domain/service"),
 	}
 }
 
 func (r *ReminderImpl) Add(ctx context.Context, item *model.ReminderItem) error {
-	ctx, span := r.tracer.Start(ctx, "Reminder#Add")
+	ctx, span := tracer.Start(ctx, "Reminder#Add")
 	defer span.End()
 
 	if err := r.reminder.Add(ctx, item); err != nil {
@@ -64,7 +60,7 @@ func (r *ReminderImpl) Add(ctx context.Context, item *model.ReminderItem) error 
 }
 
 func (r *ReminderImpl) List(ctx context.Context, conversationID model.ConversationID) (model.ReminderItems, error) {
-	ctx, span := r.tracer.Start(ctx, "Reminder#List")
+	ctx, span := tracer.Start(ctx, "Reminder#List")
 	defer span.End()
 
 	items, err := r.reminder.List(ctx, conversationID)
@@ -75,7 +71,7 @@ func (r *ReminderImpl) List(ctx context.Context, conversationID model.Conversati
 }
 
 func (r *ReminderImpl) Get(ctx context.Context, conversationID model.ConversationID, itemID model.ReminderItemID) (*model.ReminderItem, error) {
-	ctx, span := r.tracer.Start(ctx, "Reminder#Get")
+	ctx, span := tracer.Start(ctx, "Reminder#Get")
 	defer span.End()
 
 	item, err := r.reminder.Get(ctx, conversationID, itemID)
@@ -86,7 +82,7 @@ func (r *ReminderImpl) Get(ctx context.Context, conversationID model.Conversatio
 }
 
 func (r *ReminderImpl) Delete(ctx context.Context, conversationID model.ConversationID, itemID model.ReminderItemID) error {
-	ctx, span := r.tracer.Start(ctx, "Reminder#Delete")
+	ctx, span := tracer.Start(ctx, "Reminder#Delete")
 	defer span.End()
 
 	item, err := r.reminder.Get(ctx, conversationID, itemID)
@@ -109,7 +105,7 @@ func (r *ReminderImpl) Delete(ctx context.Context, conversationID model.Conversa
 }
 
 func (r *ReminderImpl) ListAll(ctx context.Context) (model.ReminderItems, error) {
-	ctx, span := r.tracer.Start(ctx, "Reminder#ListAll")
+	ctx, span := tracer.Start(ctx, "Reminder#ListAll")
 	defer span.End()
 
 	items, err := r.reminder.ListAll(ctx)
@@ -120,7 +116,7 @@ func (r *ReminderImpl) ListAll(ctx context.Context) (model.ReminderItems, error)
 }
 
 func (r *ReminderImpl) SyncSchedule(ctx context.Context, items model.ReminderItems) error {
-	ctx, span := r.tracer.Start(ctx, "Reminder#SyncSchedule")
+	ctx, span := tracer.Start(ctx, "Reminder#SyncSchedule")
 	defer span.End()
 
 	now := time.Now()
