+++ presentation/http/middleware.go
@@ -3,9 +3,6 @@ package http
 import (
 	"net/http"
 
-	"github.com/GoogleCloudPlatform/opentelemetry-operations-go/propagator"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-	"go.opentelemetry.io/otel/propagation"
 	"go.uber.org/zap"
 
 	"github.com/ww24/linebot/logger"
@@ -23,20 +20,5 @@ func PanicHandler(log *logger.Logger) func(http.Handler) http.Handler {
 
 			next.ServeHTTP(w, r)
 		})
-	}
-}
-
-func XCTCOpenTelemetry() func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return otelhttp.NewHandler(next, "server",
-			otelhttp.WithMessageEvents(otelhttp.ReadEvents, otelhttp.WriteEvents),
-			otelhttp.WithPropagators(
-				propagation.NewCompositeTextMapPropagator(
-					propagator.CloudTraceOneWayPropagator{},
-					propagation.TraceContext{},
-					propagation.Baggage{},
-				),
-			),
-		)
 	}
 }
